Crate sdf

Read .sdf files, full waveform data files from Riegl Laser Measurement Systems.

This is in part a wrapper library around libsdfifc.so, Riegl's sdf library. This library also provides functions to convert .sdf files to discrete return .sdc files using Gaussian decomposition.

The sdfifc library is not thread-safe, and so this library should only be used in single-threaded applications. When running this library's test suite, you must set RUST_TEST_THREADS=1 or else you most likely will get a segfault.

Reexports

pub use error::Error;

pub use file::File;

Modules

convert
Convert sdf files to other formats.

error
Wrapper around fwifc's error handling routines.

file
Public, safe wrappers around fwifc_file and its member functions.

Structs

LibraryVersion
Container structure for information about the library.

Functions

library_version
Returns information about the fwifc library.

Type Definitions

Result
Our custom result type.