Struct sdf::convert::Point

pub struct Point {
    pub time: f64,
    pub range: f32,
    pub theta: f32,
    pub x: f32,
    pub y: f32,
    pub z: f32,
    pub target: u8,
    pub num_target: u8,
    pub facet: u16,
    pub peak: Peak<u16>,
    pub high_channel: bool,
}

A 3D point in the scanner's own coordiante frame.

Fields

time

The time that this point was collected. Its reference frame depends on the time settings provided to the sdf library, which at this point are poorly defined.

range

The raw range to the point, which is different than the cartesian xyz distance.

theta

The mirror scan angle in degrees.

x

The x coordinate of the point in the scanner's own coordinate system.

y

The y coordinate of the point in the scanner's own coordinate system.

z

The z coordinate of the point in the scanner's own coordinate system.

target

The target number (1-indexed).

num_target

The total number of targets in this pulse.

facet

The mirror facet used to reflect the laser energy.

peak

The raw peak information returned from peakbag.

high_channel

Was this point collected on the high channel?